Move batch operations to standalone directories under each language
