+++ java/batch_operations/batch_test_setup.sql
