+++ java/batch_operations/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'dsql-batch-operations'
