+++ java/batch_operations/src/main/java/com/example/dsql/batch_operations/BatchDelete.java
