+++ javascript/batch_operations/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "dsql-batch-operations",
+  "version": "1.0.0",
+  "description": "Batch DELETE and UPDATE operations for Aurora DSQL",
+  "main": "src/main.js",
+  "scripts": {
+    "start": "node src/main.js",
+    "test": "jest --runInBand --detectOpenHandles --forceExit"
+  },
+  "dependencies": {
+    "@aws/aurora-dsql-node-postgres-connector": "^0.1.8",
+    "pg": "^8.13.0"
+  },
+  "devDependencies": {
+    "jest": "^30.3.0"
+  },
+  "license": "MIT-0"
+}
