+++ javascript/batch_operations/src/batchDelete.js
