+++ javascript/batch_operations/src/batchUpdate.js
