+++ javascript/batch_operations/test/batch_operations.test.js
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
 
-const { main } = require("../../src/batch_operations/main");
+const { main } = require("../src/main");
 
 test("Batch operations demo runs successfully", async () => {
   await main();
