+++ python/batch_operations/src/batch_delete.py
